hg clone: don't try to delete "."
-----BEGIN PGP SIGNED MESSAGE-----
Hash: SHA1

hg clone: don't try to delete "."

manifest hash: 7e7bd5375f7729812b7188b7a53287b3c6049823
-----BEGIN PGP SIGNATURE-----
Version: GnuPG v1.4.0 (GNU/Linux)

iD8DBQFCwZvlywK+sNU5EO8RAjn6AKCaBNWdLSNXIDqQZ3vMcN5ie5ZZaQCfaG37
hAN32PbYoQh8UC88AeOKOk4=
=kBOD
-----END PGP SIGNATURE-----

+++ mercurial/commands.py
@@ -269,7 +269,7 @@ def clone(ui, source, dest = None, **opts):
     """make a copy of an existing repository"""
     source = ui.expandpath(source)
 
-    created = success = False
+    success = False
 
     if dest is None:
         dest = os.getcwd()
@@ -314,7 +314,7 @@ def clone(ui, source, dest = None, **opts):
         success = True
         
     finally:
-        if not success:
+        if created and not success:
             import shutil
             shutil.rmtree(dest, True)
